+++ src/utils.ts
@@ -1,17 +1,38 @@
-import { VarNode, Node } from './node';
+import { Node, SoftNode } from './node';
 
+/* js version for testing
+function clone(originalObj) {
+  const newObj = Array.isArray(originalObj) ? [] : {};
+  for (const [key, value] of Object.entries(originalObj))
+    newObj[key] = (typeof value === 'object' && value !== null) ? clone(value) : value;
+
+  return newObj;
+}
+*/
 
 // https://stackoverflow.com/a/34624648/10247962
-export function deepClone<T>(originalObj: T): T {
-  const newObj: any = Array.isArray(originalObj) ? [] : {};
-  for (const [key, value] of Object.entries(originalObj))
-    newObj[key] = (typeof value === 'object' && value !== null) ? deepClone(value) : value;
+export function deepClone<T extends obj>(obj: T): T {
+  const newObj: any = Array.isArray(obj) ? [] : {};
+  for (const [key, value] of Object.entries(obj))
+    newObj[key] = isObject(value) ? deepClone(value) : value;
 
   return newObj;
 }
 
+/**
+ * Will deep clone only the children of the object that are Nodes. Else, will only shallow copy.
+ * Did this because using deep clone in a Model with database set (in _root() or in _clone())
+ * would have an infinite recursion, as Database points to itself.
+*/
+export function deepCloneNode<T extends SoftNode>(model: T): T {
+  const newObj: any = {};
+  for (const [key, value] of Object.entries(model))
+    newObj[key] = isNode(value) ? deepCloneNode(value) : value;
 
-export type obj = Record<string, unknown>;
+  return newObj;
+}
+
+export type obj<T = unknown> = Record<string, T>;
 
 // Returns if data is a non-null object
 export function isObject(data: any): data is obj {
@@ -37,9 +58,35 @@ export function isVarNode(data: any): boolean {
 }
 
 // If data is a Node and have a VarNode as child, return this VarNode obj. Else, null.
-export function getVarNodeChild(data: any): VarNode | undefined {
-  if (!isObject(data))
-    return;
-  return Object.values(data).find((e: any) => isVarNode(e)) as VarNode;
+export function getVarNodeChildKey(childrenData: any): string | undefined {
+  if (!isObject(childrenData))
+    return undefined;
+  // Should have just one child.
+  return Object.entries(childrenData).find(([key, data]) => isVarNode(data))?.[0];
+}
+
+
+// _varNodeChildKey in params to type check the model already have one.
+// Also keeps the function safe if the _varNodeChildKey property changes.
+// (as the "as any" would accept an old term not being used anymore)
+export function getVarNodeChild(model: SoftNode, _varNodeChildKey: string): SoftNode {
+  return (model as any)[_varNodeChildKey];
+}
+
+export function getNodeChild(model: SoftNode, _nodeChildKey: string): SoftNode {
+  return (model as any)[_nodeChildKey];
+}
+
+
+
+export function getNodeChildrenKeys(children: any): string[] {
+  if (!isObject(children))
+    return [];
+  const result: string[] = [];
+  for (const [key, child] of Object.entries(children))
+    if (isNode(child))
+      result.push(key);
+
+  return result;
 }